+++ InventoryManagement/SmartPickup/PreviousSlot.cs
@@ -34,7 +34,7 @@ public sealed class PreviousSlot : ModSystem {
     private static void HookMarkOnLeftClick(On_ItemSlot.orig_LeftClick_ItemArray_int_int orig, Item[] inv, int context, int slot) => UpdateMark((inv, context, slot) => orig(inv, context, slot), inv, context, slot, Main.mouseLeft && Main.mouseLeftRelease);
     private static void HookMarkOnRightClick(On_ItemSlot.orig_RightClick_ItemArray_int_int orig, Item[] inv, int context, int slot) => UpdateMark((inv, context, slot) => orig(inv, context, slot), inv, context, slot, Main.mouseRight);
     private static void UpdateMark(Action<Item[], int, int> orig, Item[] inv, int context, int slot, bool update) {
-        if (!update || !Configs.PreviousSlot.Mouse || !InventoryLoader.IsInventorySlot(Main.LocalPlayer, inv, context, slot, out Slot mark)) {
+        if (!update || !Configs.PreviousSlot.Mouse || !InventoryLoader.IsInventorySlot(Main.LocalPlayer, inv, context, slot, out InventorySlot mark)) {
             orig(inv, context, slot);
             return;
         }
@@ -42,8 +42,8 @@ public sealed class PreviousSlot : ModSystem {
         orig(inv, context, slot);
         if (Main.mouseItem.type == oldMouse) return;
 
-        bool removed = oldType != ItemID.None && (Configs.SmartPickup.Value.previousSlot == Configs.ItemPickupLevel.AllItems || mark.Inventory.CanBePrimary || oldFav);
-        bool placed = inv[slot].type != ItemID.None && (Configs.SmartPickup.Value.previousSlot == Configs.ItemPickupLevel.AllItems || mark.Inventory.CanBePrimary || inv[slot].favorited);
+        bool removed = oldType != ItemID.None && (Configs.SmartPickup.Value.previousSlot == Configs.ItemPickupLevel.AllItems || mark.Inventory.CanBePreferredInventory || oldFav);
+        bool placed = inv[slot].type != ItemID.None && (Configs.SmartPickup.Value.previousSlot == Configs.ItemPickupLevel.AllItems || mark.Inventory.CanBePreferredInventory || inv[slot].favorited);
 
         if (placed && removed) Remark(inv[slot].type, oldType, oldFav);
         else if (removed) Mark(oldType, mark, oldFav);
@@ -57,10 +57,10 @@ public sealed class PreviousSlot : ModSystem {
             return;
         }
 
-        foreach (ModSubInventory inventory in InventoryLoader.SubInventories) {
-            IList<Item> items = inventory.Items(self);
+        foreach (ModSubInventory inventory in InventoryLoader.GetInventories(self)) {
+            IList<Item> items = inventory.Items;
             for (int i = 0; i < items.Count; i++) {
-                if (!items[i].IsAir && (Configs.SmartPickup.Value.previousSlot == Configs.ItemPickupLevel.AllItems || inventory.CanBePrimary || items[i].favorited)) Mark(items[i].type, new(inventory, i), items[i].favorited);
+                if (!items[i].IsAir && (Configs.SmartPickup.Value.previousSlot == Configs.ItemPickupLevel.AllItems || inventory.CanBePreferredInventory || items[i].favorited)) Mark(items[i].type, new(inventory, i), items[i].favorited);
             }
         }
         orig(self);
@@ -69,9 +69,9 @@ public sealed class PreviousSlot : ModSystem {
     public static Item PickupItemToPreviousSlot(Player player, Item item, GetItemSettings settings) {
         if (player.whoAmI != Main.myPlayer) return item;
 
-        List<Slot> slots = new();
-        while (ConsumeMark(item.type, out (Slot slot, bool favorited) mark)) {
-            IList<Item> items = mark.slot.Inventory.Items(player);
+        List<InventorySlot> slots = new();
+        while (ConsumeMark(item.type, out (InventorySlot slot, bool favorited) mark)) {
+            IList<Item> items = mark.slot.Inventory.Items;
             if (mark.slot.Index >= items.Count) continue;
             if (!(item.favorited || mark.favorited) && items[mark.slot.Index].favorited) continue;
 
@@ -79,30 +79,30 @@ public sealed class PreviousSlot : ModSystem {
             (Item moved, items[mark.slot.Index]) = (items[mark.slot.Index], new());
             Item toMove = item.Clone();
             toMove.stack = 1;
-            if (mark.slot.GetItem(player, toMove, settings).IsAir) item.stack--;
-            moved = mark.slot.GetItem(player, moved, settings);
+            if (mark.slot.GetItem(toMove, settings).IsAir) item.stack--;
+            moved = mark.slot.GetItem(moved, settings);
             player.GetDropItem(ref moved);
             if (item.IsAir) return item;
             slots.Add(mark.slot);
         }
-        foreach (Slot slot in slots) {
-            item = slot.GetItem(player, item, settings);
+        foreach (InventorySlot slot in slots) {
+            item = slot.GetItem(item, settings);
             if (item.IsAir) return item;
         }
         return item;
     }
 
     public static bool IsMarked(int type) => s_marks.TryGetValue(type, out var marks) && marks.Count > 0;
-    public static bool IsMarked(Slot slot) => s_marksData.ContainsKey(slot);
-    public static bool TryGetMark(Slot slot, [MaybeNullWhen(false)] out Item mark) => s_marksData.TryGetValue(slot, out mark);
-    public static bool ConsumeMark(int type, [MaybeNullWhen(false)] out (Slot slot, bool favorited) mark) {
+    public static bool IsMarked(InventorySlot slot) => s_marksData.ContainsKey(slot);
+    public static bool TryGetMark(InventorySlot slot, [MaybeNullWhen(false)] out Item mark) => s_marksData.TryGetValue(slot, out mark);
+    public static bool ConsumeMark(int type, [MaybeNullWhen(false)] out (InventorySlot slot, bool favorited) mark) {
         Queue<(int type, int depth)> items = [];
         int checksLeft = Configs.PreviousSlot.Value.materials ? Configs.PreviousSlot.Value.materials.Value.maxChecks : 1;
         items.Enqueue((type, Configs.PreviousSlot.Value.materials.Value.maxDepth));
         HashSet<int> added = [type];
         while (items.TryDequeue(out var item)) {
             if (IsMarked(item.type)) {
-                Slot slot = s_marks[item.type][^1];
+                InventorySlot slot = s_marks[item.type][^1];
                 mark = (slot, s_marksData[slot].favorited);
                 Unmark(slot);
                 return true;
@@ -121,7 +121,7 @@ public sealed class PreviousSlot : ModSystem {
         return false;
     }
 
-    public static void Mark(int type, Slot slot, bool favorited) {
+    public static void Mark(int type, InventorySlot slot, bool favorited) {
         if (IsMarked(slot)) {
             if (!Configs.PreviousSlot.Value.overridePrevious) return;
             Unmark(slot);
@@ -132,10 +132,10 @@ public sealed class PreviousSlot : ModSystem {
     }
     public static void Unmark(int type) {
         if (!IsMarked(type)) return;
-        foreach (Slot mark in s_marks[type]) s_marksData.Remove(mark);
+        foreach (InventorySlot mark in s_marks[type]) s_marksData.Remove(mark);
         s_marks.Remove(type);
     }
-    public static void Unmark(Slot slot) {
+    public static void Unmark(InventorySlot slot) {
         if (!IsMarked(slot)) return;
         s_marks[s_marksData[slot].type].Remove(slot);
         s_marksData.Remove(slot);
@@ -143,7 +143,7 @@ public sealed class PreviousSlot : ModSystem {
     public static void Remark(int oldType, int newType, bool? favorited = null) {
         Unmark(newType);
         if (!IsMarked(oldType)) return;
-        foreach (Slot slot in s_marks[oldType]) s_marksData[slot] = new(newType) { favorited = favorited ?? s_marksData[slot].favorited };
+        foreach (InventorySlot slot in s_marks[oldType]) s_marksData[slot] = new(newType) { favorited = favorited ?? s_marksData[slot].favorited };
         s_marks[newType] = s_marks[oldType];
         s_marks.Remove(oldType);
     }
@@ -213,12 +213,12 @@ public sealed class PreviousSlot : ModSystem {
     private static bool s_ilBackgroundMark;
 
     private static bool TryDrawMark(SpriteBatch spriteBatch, Item[] inv, int context, int slot, Vector2 position, float scale, Texture2D texture, Color color, Configs.IPreviousDisplay ui) {
-        if (!InventoryLoader.IsInventorySlot(Main.LocalPlayer, inv, context, slot, out Slot itemSlot) || !TryGetMark(itemSlot, out Item? mark)) return false;
+        if (!InventoryLoader.IsInventorySlot(Main.LocalPlayer, inv, context, slot, out InventorySlot itemSlot) || !TryGetMark(itemSlot, out Item? mark)) return false;
         ItemSlot.DrawItemIcon(mark, ItemSlot.Context.InWorld, spriteBatch, position + texture.Size() * ui.position * scale, scale * ui.scale, 32f, color * Main.cursorAlpha * ui.intensity);
         return true;
     }
 
-    private static readonly Dictionary<Slot, Item> s_marksData = [];
-    private static readonly Dictionary<int, List<Slot>> s_marks = [];
+    private static readonly Dictionary<InventorySlot, Item> s_marksData = [];
+    private static readonly Dictionary<int, List<InventorySlot>> s_marks = [];
 
 }